ARM support

+++ include/tsp.h
@@ -1,6 +1,7 @@
 #ifndef KMCOMP_TSP_H
 #define KMCOMP_TSP_H
 
+#include <cstdint>
 #include <vptree.h>
 
 namespace kmcomp
@@ -15,10 +16,10 @@ namespace kmcomp
     std::size_t build_double_ended_NN(const char* const MATRIX, const std::size_t COLUMNS, const std::size_t SUBSAMPLED_ROWS, const std::size_t OFFSET, std::vector<std::uint64_t>& order);
 
     //Distance computation between two columns
-    double columns_hamming_distance(const char* const MATRIX, const std::size_t NB_ROWS, const std::size_t COLUMN_A, const std::size_t COLUMN_B);
+    double columns_hamming_distance(const char* const MATRIX, const std::size_t NB_ROWS, const std::uint64_t COLUMN_A, const std::uint64_t COLUMN_B);
 
     //Get Nearest-Neighbor using VPTree
-    IndexDistance find_closest_vertex(VPTree<std::uint64_t>& VPTREE, const std::size_t VERTEX, const std::vector<bool>& ALREADY_ADDED);
+    IndexDistance find_closest_vertex(VPTree<std::uint64_t>& VPTREE, const std::uint64_t VERTEX, const std::vector<bool>& ALREADY_ADDED);
 
     //Hamming distance between two buffers
     std::size_t hamming_distance(const std::uint8_t* const BUFFER1, const std::uint8_t* const BUFFER2, const std::size_t LENGTH);
